Praying to god
