+++ Makefile
@@ -50,6 +50,7 @@ SRC 		:= 	main/minishell.c 				\
 				executor/executor_utils6.c		\
 				executor/executor_utils7.c		\
 				executor/executor_utils8.c		\
+				executor/executor_utils9.c		\
 				executor/piping.c 				\
 				executor/piping_utils1.c		\
 				executor/redirection.c 			\
