+++ src/executor/executor_utils3.c
@@ -1,117 +1,120 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   executor_utils3.c                                  :+:      :+:    :+:   */
+/*   executor_utils2.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: hmunoz-g <hmunoz-g@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/25 11:42:26 by hmunoz-g          #+#    #+#             */
-/*   Updated: 2025/01/10 20:59:43 by hmunoz-g         ###   ########.fr       */
+/*   Updated: 2025/01/10 21:17:40 by hmunoz-g         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/minishell.h"
 
 /*
-Checks if the current command includes a heredoc ('<<').
+Executes a single command in the Minishell.
 Steps:
-  1. Iterates through the cmd_args array to find the '<<' token.
-  2. Returns 1 if a heredoc is present, 0 otherwise.
-TODO: check return values for consistency, may need to be switched.
+  1. Validates the cmd_args array for memory allocation.
+  2. Handles redirection if required:
+      - If redirection fails, exits for non-builtin commands.
+  3. Checks if the command is a builtin:
+      - Executes the builtin and handles rerouting if necessary.
+  4. For non-builtin commands:
+      - Delegates execution to ms_handle_system_cmd.
+      - Cleans up memory if system command handling fails.
+Returns:
+  - 1 on success or handled error.
+  - Exits the process on fatal errors during redirection or command execution.
 */
-int	ms_has_heredoc(t_ms *ms)
+int	ms_exec_command(t_ms *ms, char **env)
 {
-	int	i;
-	int	hd;
-
-	i = -1;
-	hd = 0;
-	while (ms->cmd_args[++i])
+	if (!ms->cmd_args)
+		return (ms_error_handler(ms, "Error: Mem alloc failed", 1), 1);
+	if (ms_has_redirection(ms))
 	{
-		if (!ft_strncmp(ms->cmd_args[i], "<<", 2))
-			hd++;
-	}
-	return (hd);
-}
-
-/*
-Processes the heredoc setup for a command.
-Steps:
-  1. Searches for the '<<' token in cmd_args to identify the delimiter.
-  2. If no delimiter is found, handles the error and returns -1.
-  3. Calls ms_handle_heredoc to process the heredoc and obtain the FD.
-  4. Updates heredoc_fd in the Minishell state if successful.
-  5. Returns 0 on success, -1 on failure.
-*/
-int	ms_process_heredoc(t_ms *ms)
-{
-	int		fd;
-	char	*delimiter;
-	int		i;
-	int		result;
-
-	delimiter = NULL;
-	i = -1;
-	while (ms->cmd_args[++i])
-	{
-		if (!ft_strncmp(ms->cmd_args[i], "<<", 2))
+		if (ms_redirection(ms) == -1)
 		{
-			delimiter = ms->cmd_args[i + 1];
+			if (ms_is_builtin(ms->filt_args[0]))
+				return (1);
+			exit(1);
 		}
 	}
-	if (!delimiter)
-		return (ms_handle_heredoc_error(ms, "Heredoc delimiter missing"));
-	result = ms_handle_heredoc(delimiter, &fd);
-	if (result == -1)
-		return (ms_handle_heredoc_error(ms, "Heredoc processing failed"));
-	ms->heredoc_fd = fd;
+	if (ms_is_builtin(ms->cmd_args[0]))
+	{
+		if (ms_reroute_builtins(ms, env))
+			return (1);
+	}
+	else if (ms_handle_system_cmd(ms, env) == -1)
+	{
+		ft_free(ms->cmd_args);
+		ft_free(ms->filt_args);
+		return (1);
+	}
 	return (0);
 }
 
 /*
-Sets up the environment for heredoc processing.
+Handles the child process execution logic for a Minishell command.
+Received int i is the current exec chunk index in the exec pipeline.
 Steps:
-  1. Saves the current stdin and stdout file descriptors.
-  2. Redirects stdin and stdout to /dev/tty for interactive input.
-  3. Calls ms_process_heredoc to handle the heredoc processing.
-  4. Restores the original stdin and stdout file descriptors.
-  5. Cleans up resources and returns the result of heredoc processing.
-  6. Returns 0 on success, -1 on failure.
+  1. If a heredoc is present, duplicates its file descriptor to STDIN_FILENO.
+  2. Sets up the appropriate pipes for the child process.
+  3. Handles redirection unless a heredoc is present.
+  4. Executes the command (builtin or system).
+  5. Exits the child process with a failure status if any error occurs.
 */
-int	ms_handle_heredoc_setup(t_ms *ms)
+int	ms_handle_child_process(t_ms *ms, char **env, int i)
 {
-	int	orig_stdin;
-	int	orig_stdout;
-	int	result;
-	int	tty;
-
-	orig_stdin = dup(STDIN_FILENO);
-	orig_stdout = dup(STDOUT_FILENO);
-	tty = open("/dev/tty", O_RDWR);
-	if (tty != -1)
+	ms_setup_child_pipes(ms, i, ms->pipe_count);
+	if (i == 0 && ms->heredoc_fd != -1)
 	{
-		dup2(tty, STDIN_FILENO);
-		dup2(tty, STDOUT_FILENO);
-		close(tty);
+		if (dup2(ms->heredoc_fd, STDIN_FILENO) == -1)
+		{
+			perror("heredoc dup2 failed");
+			ms_error_handler(ms, "Heredoc dup2 failed", 0);
+			exit(1);
+		}
+		close(ms->heredoc_fd);
 	}
-	result = ms_process_heredoc(ms);
-	dup2(orig_stdin, STDIN_FILENO);
-	dup2(orig_stdout, STDOUT_FILENO);
-	close(orig_stdin);
-	close(orig_stdout);
-	return (result);
+	else if (i > 0)
+		ms->heredoc_fd = -1;
+	if (ms_has_redirection(ms))
+	{
+		if (ms_redirection(ms) == -1)
+			exit(1);
+	}
+	if (ms_exec_command(ms, env) != 0)
+		exit(1);
+	exit(127);
 }
 
 /*
-Handles errors encountered during heredoc setup or processing.
+Handles the parent process logic for Minishell.
 Steps:
-  1. Checks the current shell state using ms_get_set.
-  2. Logs the provided error message if the shell state is 2 (heredoc input).
-  3. Returns -1 to indicate failure.
+  1. Closes the heredoc file descriptor if it was used.
+  2. Resets the heredoc_fd in the Minishell state to indicate closure.
 */
-int	ms_handle_heredoc_error(t_ms *ms, char *error_msg)
+int	ms_handle_parent_process(t_ms *ms)
 {
-	if (ms_get_set(GET, 0) == 2)
-		ms_error_handler(ms, error_msg, 0);
-	return (-1);
+	if (ms->heredoc_fd != -1)
+	{
+		close(ms->heredoc_fd);
+		ms->heredoc_fd = -1;
+	}
+	return (0);
+}
+
+/*
+Closes the appropriate pipe file descriptors after use.
+Steps:
+  1. Closes the read end of the previous pipe (if applicable).
+  2. Closes the write end of the current pipe.
+*/
+void	ms_close_used_pipes(int **pipe_fds, int i)
+{
+	if (i > 0)
+		close(pipe_fds[i - 1][0]);
+	if (pipe_fds[i])
+		close(pipe_fds[i][1]);
 }
