+++ src/executor/executor_utils8.c
@@ -1,78 +1,98 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   executor_utils8.c                                  :+:      :+:    :+:   */
+/*   executor_utils7.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: hmunoz-g <hmunoz-g@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/25 11:42:26 by hmunoz-g          #+#    #+#             */
-/*   Updated: 2025/01/09 10:52:40 by hmunoz-g         ###   ########.fr       */
+/*   Updated: 2025/01/14 08:34:58 by hmunoz-g         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/minishell.h"
 
-int	ms_detect_space_arg(const char *chunk)
-{
-	unsigned long	i;
-	int				count;
-
-	i = 0;
-	count = 0;
-	while (chunk[i] && i < ft_strlen(chunk) - 2)
-	{
-		if (ft_isspace(chunk[i]) && ft_isspace(chunk[i + 1])
-			&& ft_isspace(chunk[i + 2]))
-			count++;
-		i ++;
-	}
-	return (count);
-}
-
 /*
-Helper function to handle space arguments (e.g.: tr " " "x").
-It's the first of a two step process, in which spaces get 
-	temporarily changed to " so that tokenization don't skip them. 
+Cleans up the executor env by clearing lists and freeing allocated memory.
+Steps:
+  1. Clears the list of `chain_tokens`, freeing each element.
+  2. Frees the `exec_chunks` array and the environment array.
 */
-char	*ms_process_space_args_in(char *chunk)
-{
-	unsigned long	i;
-
-	i = 1;
-	while (chunk[i] && i < ft_strlen(chunk) - 2)
-	{
-		if (ft_isspace(chunk[i - 1]) && ft_isspace(chunk[i])
-			&& ft_isspace(chunk[i + 1]))
-			chunk[i] = '"';
-		i++;
-	}
-	return (chunk);
-}
-
-/*
-Helper function to handle space arguments (e.g.: tr " " "x").
-It's the second of a two step process, in which temporarily
-	changed tokens (") get changed back to a space char.
-*/
-char	**ms_process_space_args_out(char **args)
+void	ms_executor_cleanup(t_ms *ms, char **env)
 {
 	int	i;
 
-	i = -1;
-	while (args[++i])
+	i = 0;
+	while (i < ms->chunk_count)
 	{
-		if (args[i][0] == '"')
-			args[i][0] = ' ';
+		ms_tokclear(&ms->exec_chunks[i], free);
+		i++;
 	}
-	return (args);
+	free(ms->exec_chunks);
+	ms->exec_chunks = NULL;
+	ft_free(env);
 }
 
-int	ms_heredoc_interruption(t_ms *ms, char **env)
+/*
+Cleans up arguments by freeing allocated memory.
+Steps:
+  1. Checks if `filt_args` is not NULL, then frees it.
+  2. Checks if `cmd_args` is not NULL, then frees it.
+*/
+void	ms_cleanup_args(t_ms *ms)
 {
-	ms_get_set(SET, SHELL_NORMAL);
-	ms_executor_cleanup(ms, env);
-	ms_free_pipes(ms->pipe_fds, ms->pipe_count);
-	ms_cleanup_heredoc(ms);
-	ms->exit_status = 130;
-	return (130);
+	if (ms->filt_args[0] != NULL)
+	{
+		free(ms->filt_args);
+		ms->filt_args = NULL;
+	}
+	if (ms->cmd_args[0] != NULL)
+	{
+		free(ms->cmd_args);
+		ms->cmd_args = NULL;
+	}
+}
+
+/*
+Cleans up heredoc file descriptor.
+Steps:
+  1. Checks if the heredoc file descriptor (heredoc_fd) is valid (not -1).
+  2. If valid, closes the file descriptor and resets it to -1.
+*/
+void	ms_cleanup_heredoc(t_ms *ms)
+{
+	if (ms->heredoc_fd != -1)
+	{
+		close(ms->heredoc_fd);
+		ms->heredoc_fd = -1;
+	}
+}
+
+/*
+Saves the current standard input, output, and error file descriptors.
+Steps:
+  1. Duplicates the current standard file descriptors.
+  2. Saves STD FDS in the saved_fds array.
+*/
+void	ms_save_std_fds(int *saved_fds)
+{
+	saved_fds[0] = dup(STDIN_FILENO);
+	saved_fds[1] = dup(STDOUT_FILENO);
+	saved_fds[2] = dup(STDERR_FILENO);
+}
+
+/*
+Restores the previously saved standard input, output, and error file descriptors.
+Steps:
+  1. Restores the file descriptors from the saved_fds array using dup2().
+  2. Closes the saved file descriptors.
+*/
+void	ms_restore_std_fds(int *saved_fds)
+{
+	dup2(saved_fds[0], STDIN_FILENO);
+	dup2(saved_fds[1], STDOUT_FILENO);
+	dup2(saved_fds[2], STDERR_FILENO);
+	close(saved_fds[0]);
+	close(saved_fds[1]);
+	close(saved_fds[2]);
 }
